requirements.txt
numpy
pyyaml
scholarly
rendercv[full]
free-proxy
httpx==0.23.3
